add tests and implement WeatherRepositoryImpl

+++ Android/features/weather/src/main/java/com/kkunan/weather/data/models/GetWeatherByLatLng.kt
@@ -6,34 +6,34 @@ sealed class GetWeatherByLatLng {
     data class Response(
         val base: String? = null,
         val clouds: Clouds? = null,
-        val cod: Int? = null,
+        val cod: Double? = null,
         val coord: Coord? = null,
-        val dt: Int? = null,
-        val id: Int? = null,
+        val dt: Double? = null,
+        val id: Double? = null,
         val main: Main? = null,
         val name: String? = null,
         val sys: Sys? = null,
-        val timezone: Int? = null,
-        val visibility: Int? = null,
+        val timezone: Double? = null,
+        val visibility: Double? = null,
         val weather: List<Weather>? = null,
         val wind: Wind? = null
     )
 
     data class Clouds(
-        val all: Int? = null
+        val all: Double? = null
     )
 
     data class Coord(
-        val lat: Int? = null,
-        val lon: Int? = null
+        val lat: Double? = null,
+        val lon: Double? = null
     )
 
     data class Main(
         val feels_like: Double? = null,
-        val grnd_level: Int? = null,
-        val humidity: Int? = null,
-        val pressure: Int? = null,
-        val sea_level: Int? = null,
+        val grnd_level: Double? = null,
+        val humidity: Double? = null,
+        val pressure: Double? = null,
+        val sea_level: Double? = null,
         val temp: Double? = null,
         val temp_max: Double? = null,
         val temp_min: Double? = null
@@ -41,19 +41,19 @@ sealed class GetWeatherByLatLng {
 
     data class Sys(
         val country: String? = null,
-        val sunrise: Int? = null,
-        val sunset: Int? = null
+        val sunrise: Double? = null,
+        val sunset: Double? = null
     )
 
     data class Weather(
         val description: String? = null,
         val icon: String? = null,
-        val id: Int? = null,
+        val id: Double? = null,
         val main: String? = null
     )
 
     data class Wind(
-        val deg: Int? = null,
+        val deg: Double? = null,
         val gust: Double? = null,
         val speed: Double? = null
     )
